refactor: move lazy loc/range machinery to lib/javascript/syntax.js
Match the syntax.js layout used by lib/css and lib/html.
